requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==20.1.0
chardet==3.0.4
idna==2.10
idna-ssl==1.1.0
multidict==4.7.6
pendulum==2.1.2
pymongo==3.11.0
python-dateutil==2.8.1
pytzdata==2020.1
six==1.15.0
slackclient==2.8.0
typing-extensions==3.7.4.3
yarl==1.5.1
